pyproject.toml
[project]
name = "ipygame"
version = "0.1.0"
description = "A pygame compatibility layer for Jupyter notebooks using ipycanvas"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "ipycanvas>=0.13",
    "ipywidgets>=8.0",
    "numpy>=1.24",
    "Pillow>=9.0",
]

[dependency-groups]
dev = ["pydoc-markdown>=4.8.2", "PyYAML>=6.0"]

[project.urls]
Homepage = "https://github.com/Kamuyin/ipygame"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
